fix(trusty-search): split DOC-37 repo_identity tests out of tests_list.rs (line-cap)
The fix pass added two DOC-37 tests + two helpers to tests_list.rs, pushing it
to 501 SLOC — one over the 500-SLOC production cap (tests_list.rs does not
qualify for the 1500-SLOC test-file cap: it ends in `_list.rs`, not
`_tests.rs`/`tests.rs`, per the classification rule in CLAUDE.md /
check_line_cap.sh). `bash scripts/check_line_cap.sh` failed with "new
oversized file" — the ratchet forbids adding one.

Moved `list_indexes_repo_identity_details_and_filter` and
`list_indexes_repo_identity_tree_filter` (plus the shared
`isolate_and_seed_toml` / `clear_data_dir_env` helpers) into a new sibling
`list_repo_identity_tests.rs`, whose basename ends in `_tests.rs` and so gets
the 1500-SLOC test-file cap. Registered `mod list_repo_identity_tests;`
(`#[cfg(test)]`) in `mod.rs`. tests_list.rs is back to 419 SLOC;
list_repo_identity_tests.rs is 237 SLOC — both comfortably under their caps.
Updated the two `Test:` doc-comment pointers in indexes.rs / router.rs to cite
the new file.

`bash scripts/check_line_cap.sh` now passes: 0 violations.
